+++ instrumentation/lettuce/lettuce-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/lettuce/v5_0/LettuceBatchAttributesGetter.java
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.javaagent.instrumentation.lettuce.v5_0;
 
+import io.lettuce.core.RedisURI;
 import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientAttributesGetter;
 import io.opentelemetry.semconv.incubating.DbIncubatingAttributes.DbSystemNameIncubatingValues;
 import javax.annotation.Nullable;
@@ -38,5 +39,19 @@ final class LettuceBatchAttributesGetter
   @Nullable
   public Long getDbOperationBatchSize(LettuceBatchRequest request) {
     return request.getBatchSize();
+  }
+
+  @Nullable
+  @Override
+  public String getServerAddress(LettuceBatchRequest request) {
+    RedisURI redisUri = request.getRedisUri();
+    return redisUri != null ? redisUri.getHost() : null;
+  }
+
+  @Nullable
+  @Override
+  public Integer getServerPort(LettuceBatchRequest request) {
+    RedisURI redisUri = request.getRedisUri();
+    return redisUri != null ? redisUri.getPort() : null;
   }
 }