+++ instrumentation/log4j/log4j-appender-2.17/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/log4j/appender/v2_17/Log4jHelper.java
@@ -47,6 +47,8 @@ public class Log4jHelper {
         config.getBoolean("capture_map_message_attributes/development", false);
     boolean captureMarkerAttribute =
         config.getBoolean("capture_marker_attribute/development", false);
+    boolean captureTemplate = config.getBoolean("capture_template/development", false);
+    boolean captureArguments = config.getBoolean("capture_arguments/development", false);
     List<String> captureContextDataAttributes =
         config.getScalarList("capture_mdc_attributes/development", String.class, emptyList());
     boolean v3Preview = commonConfig.getBoolean("v3_preview", false);
@@ -58,6 +60,8 @@ public class Log4jHelper {
             captureCodeAttributes,
             captureMapMessageAttributes,
             captureMarkerAttribute,
+            captureTemplate,
+            captureArguments,
             captureContextDataAttributes,
             v3Preview);
   }