+++ javaagent-tooling/src/main/java/io/opentelemetry/javaagent/tooling/AgentExtension.java
@@ -26,7 +26,19 @@ public interface AgentExtension extends Ordered {
    * @return The customized agent. Note that this method MUST return a non-null {@link AgentBuilder}
    *     instance that contains all customizations defined in this extension.
    */
-  AgentBuilder extend(AgentBuilder agentBuilder, ConfigProperties config);
+  default AgentBuilder extend(AgentBuilder agentBuilder) {
+    return agentBuilder;
+  }
+
+  /**
+   * Extend the passed {@code agentBuilder} with custom logic (e.g. instrumentation).
+   *
+   * @return The customized agent. Note that this method MUST return a non-null {@link AgentBuilder}
+   *     instance that contains all customizations defined in this extension.
+   */
+  default AgentBuilder extend(AgentBuilder agentBuilder, ConfigProperties config) {
+    return extend(agentBuilder);
+  }
 
   /**
    * Returns the name of the extension. It does not have to be unique, but it should be